Added install_extras/uninstall_extras commands to install bash completion script
Trying to install that stuff (if target location is writable) when being installed as a dependency causes a sandbox violation in easy_install which cannot be circumvented as it seems. So only install bash completion if explicitely invoked to do via "python setup.py install_extras". With "python setup.py uninstall_extras" it's also possible to uninstall.